Added KeyButton
The KeyButton can trigger samples. It shows a key label an a sample
icon. For the trigger/untrigger behavior there is a seperate class. This
can be reused for the behavior of the actual keyboard keys.
Also, a new ActivationMode was added: WHILE_TRIGGERED_ONCE. When
triggered, this will play the sample once, instead of looping it. When
untriggered, it stops (like WHILE_TRIGGERED)