Added sample recording funtionality and UI
Samples can now be recorded using the microphone. This functionality can
be controlles with the number keys and with the new buttons on top of
the keyboard. Start the recording mode, then press and hold one of the
buttons to record. Go out of recording mode and press a button with a
recorded sample to play it. In deleting mode, you can obviously delete
samples.